requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-table==5.0.0
Flask==3.0.3
gunicorn==20.0.4
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
nest-asyncio==1.6.0
numpy==2.1.2
oracledb==2.4.1
packaging==24.1
pandas==2.2.3
plotly==5.24.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
